correccion

+++ src/main/java/unlar/edu/ar/Tp3/models/Vehiculo.java
@@ -14,6 +14,6 @@ public abstract class Vehiculo {
 
     private int bateria;
 
-    private float tarifaBase;
+    private double tarifaBase;
 
 }
